Include tests in gem distribution (#5)
If someone needs to debug Gouda it helps that they won't have to look for the exact revision on GH to get the tests. And there is no sizable test corpus needed for Gouda (tests are usually excluded because of the large-ish files which are not necessary). Including tests lowers the barrier for new contributors.