Remove logger overrides from Gouda (#14)
In our app, this causes more trouble than it is worth - Gouda should use
the common Rails logger and not touch it. If we want to log less SQL
from the Gouda worker, we should look for a different way to do so - for
example, by selectively silencing the logger using blocks or by
temporarily lowering the logging level during the query.