Do not spam logs with SQL for super-frequent tasks
The whole logging configuration had the goal of not outputting too much SQL into the STDOUT of the process hosting the worker thread, primarily when using Gouda inside banksvc. But what this comes down to is that we run very frequent polling calls, which - in turn - perform SQL queries. These queries can actually be suppressed selectively.