Specify the actual Ruby version we support
We need to specify it both in the gemspec and in the standardrb config. Otherwise we use syntax features of Ruby 3.0 but we do not specify that our code actually would not load on the older Ruby version. And it is better to keep these two coordinated, so that standardrb alerts us if we use language features which our target Ruby version does not, in fact, support.