Do not execute removed cron workloads (#17)
We need to check whether a workload was enqueued with a scheduler key,
but no longer is in the cron table. If that is the case (we are trying
to execute a workload which has a scheduler key, but the scheduler does
not know about that key) it means that the workload has been removed
from the cron table and must not run. Moreover: running it can be
dangerous because it was likely removed from the table for a reason.
Should that be the case, mark the job "finished" and return `nil` to get
to the next poll. If the deployed worker still has the workload in its
scheduler table, but a new deploy removed it - this is a race condition,
but we are willing to accept it.

Note that we are already "just not enqueueing" that job when the cron
table gets loaded - but it is not enough.
If our system can be in a state of partial deployment:

```
  [  release 1 does have some_job_hourly crontab entry ]
                 [  release 2 no longer does                           ]
                 ^ --- race conditions possible here --^
```

...and we remove the crontabled workloads during app boot, it does not
give us a guarantee that release 1 won't reenqueue them. For example,
via the "reinsert next scheduled" feature when the job is executing.
This is why this safeguard is needed.

This protects us from a very dangerous failure mode where we would
remove an entry from the cron table, deploy the change, and then still
have the workload run a day later.

Also, mandate Ruby 3.1 as the minimum version and pin the Standardrb syntax to that.